# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerAuth
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegistrationInput" }
      responses:
        "201":
          description: Registered and signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "400": { description: Invalid registration data }
        "409": { description: Username already exists }
        "429": { description: Too many attempts }
        "503": { description: Registration is disabled or workspace is in maintenance mode }
  /auth/login:
    post:
      operationId: loginAuth
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "400": { description: Invalid login data }
        "401": { description: Invalid credentials }
        "429": { description: Too many attempts }
...
  /auth/me:
    get:
      operationId: getAuthUser
      tags: [auth]
      responses:
        "200":
          description: Current authenticated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "401": { description: Authentication is required }
...
components:
...
  schemas:
...
    AuthUser:
      type: object
      required: [id, username, role]
      properties:
        id: { type: string }
        username: { type: string }
        role: { type: string, enum: [user, admin] }
    RegistrationInput:
      type: object
      required: [username, password, confirmPassword]
      properties:
        username: { type: string, minLength: 3, maxLength: 32 }
        password: { type: string, minLength: 10, maxLength: 128 }
        confirmPassword: { type: string, minLength: 10, maxLength: 128 }
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 128 }
        password: { type: string, minLength: 1, maxLength: 128 }
...